+++ plugins/status/interface.go
@@ -1,4 +1,4 @@
-package checker
+package status
 
 // Status consists of status code from the service
 type Status struct {