+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 CHANGELOG
 =========
 
+v2.0.0-RC.4 (20.02.2021)
+-------------------
+
+- PHP tests use latest signatures (https://github.com/spiral/roadrunner/pull/550).
+- Endure container update to v1.0.0-RC.2 version.
+- Remove unneeded mutex from the `http.Workers` method.
+- Rename `checker` plugin package to `status`, remove `/v1` endpoint prefix (#557).
+- Add static, headers, status, gzip plugins to the `main.go`.
+- Fix workers pool behavior -> idle_ttl, ttl, max_memory are soft errors and exec_ttl is hard error.
+
 v2.0.0-RC.3 (17.02.2021)
 -------------------
 
@@ -12,7 +22,7 @@ v2.0.0-RC.3 (17.02.2021)
 v2.0.0-RC.2 (11.02.2021)
 -------------------
 
-- Update RR-Core to version v2.0.0-RC.2
+- Update RR to version v2.0.0-RC.2
 - Update Temporal plugin to version v2.0.0-RC.1
-- Update goridge to version v3.0.1
+- Update Goridge to version v3.0.1
 - Update Endure to version v1.0.0-RC.1